+++ hardhat.config.ts
@@ -1,0 +1,26 @@
+import "@nomicfoundation/hardhat-toolbox";
+import type { HardhatUserConfig } from "hardhat/config";
+
+const config: HardhatUserConfig = {
+  solidity: {
+    version: "0.8.24",
+    settings: {
+      optimizer: {
+        enabled: true,
+        runs: 200,
+      },
+      evmVersion: "cancun",
+    },
+  },
+  paths: {
+    sources: "./contracts",
+    tests: "./contracts/test",
+    cache: "./cache",
+    artifacts: "./artifacts",
+  },
+  mocha: {
+    timeout: 120000,
+  },
+};
+
+export default config;